Remove old injector
